#4 12th commit: remain power 구현
